added method add in userDao, fixed all errors

+++ src/util/TextUtil.java
@@ -39,7 +39,7 @@ public class TextUtil {
 
     }
 
-    public static void deleteFromFile(String entityName, int id){
+    public static void deleteFromFile(String entityName, long id){
 
     }
